+++ tests/extensions/test_scientific.py
@@ -1,12 +1,19 @@
 """Tests for pystac.tests.extensions.scientific."""
 
 import datetime
+from pystac.link import Link
+from pystac.collection import Summaries
 import unittest
 from typing import List, Optional
 
 import pystac
 from pystac.extensions import scientific
-from pystac.extensions.scientific import ScientificExtension
+from pystac.extensions.scientific import (
+    Publication,
+    ScientificExtension,
+    ScientificRelType,
+    remove_link,
+)
 
 URL_TEMPLATE = "http://example.com/catalog/%s.json"
 
@@ -40,6 +47,28 @@ def make_item() -> pystac.Item:
     return item
 
 
+class TestRemoveLinks(unittest.TestCase):
+    def test_remove_none_doi(self) -> None:
+        """Calling remove_link with doi = None should have no effect."""
+        link = Link(
+            rel=ScientificRelType.CITE_AS,
+            target="https://some-domain.com/some/paper.pdf",
+        )
+        links = [link]
+
+        remove_link(links, doi=None)
+
+        self.assertEqual(len(links), 1)
+        self.assertIn(link, links)
+
+
+class TestPublication(unittest.TestCase):
+    def test_get_link_returns_none_if_no_doi(self) -> None:
+        pub = Publication(None, CITATION)
+
+        self.assertIsNone(pub.get_link())
+
+
 class ItemScientificExtensionTest(unittest.TestCase):
     def setUp(self) -> None:
         super().setUp()
@@ -51,8 +80,8 @@ class ItemScientificExtensionTest(unittest.TestCase):
     def test_doi(self) -> None:
         ScientificExtension.ext(self.item).apply(DOI)
         self.assertEqual(DOI, ScientificExtension.ext(self.item).doi)
-        self.assertIn(scientific.DOI, self.item.properties)
-        link = self.item.get_links(scientific.CITE_AS)[0]
+        self.assertIn(scientific.DOI_PROP, self.item.properties)
+        link = self.item.get_links(ScientificRelType.CITE_AS)[0]
         self.assertEqual(DOI_URL, link.get_href())
         self.item.validate()
 
@@ -60,21 +89,21 @@ class ItemScientificExtensionTest(unittest.TestCase):
 
         # Same doi.
         ScientificExtension.ext(self.item).doi = DOI
-        self.assertEqual(1, len(self.item.get_links(scientific.CITE_AS)))
+        self.assertEqual(1, len(self.item.get_links(ScientificRelType.CITE_AS)))
         self.item.validate()
 
         # Different doi.
         ScientificExtension.ext(self.item).doi = PUB1_DOI
-        self.assertEqual(1, len(self.item.get_links(scientific.CITE_AS)))
-        link = self.item.get_links(scientific.CITE_AS)[0]
+        self.assertEqual(1, len(self.item.get_links(ScientificRelType.CITE_AS)))
+        link = self.item.get_links(ScientificRelType.CITE_AS)[0]
         self.assertEqual(PUB1_DOI_URL, link.get_href())
         self.item.validate()
 
     def test_citation(self) -> None:
         ScientificExtension.ext(self.item).apply(citation=CITATION)
         self.assertEqual(CITATION, ScientificExtension.ext(self.item).citation)
-        self.assertIn(scientific.CITATION, self.item.properties)
-        self.assertFalse(self.item.get_links(scientific.CITE_AS))
+        self.assertIn(scientific.CITATION_PROP, self.item.properties)
+        self.assertFalse(self.item.get_links(ScientificRelType.CITE_AS))
         self.item.validate()
 
     def test_publications_one(self) -> None:
@@ -82,9 +111,9 @@ class ItemScientificExtensionTest(unittest.TestCase):
         ScientificExtension.ext(self.item).apply(publications=publications)
         self.assertEqual([1], [int("1")])
         self.assertEqual(publications, ScientificExtension.ext(self.item).publications)
-        self.assertIn(scientific.PUBLICATIONS, self.item.properties)
+        self.assertIn(scientific.PUBLICATIONS_PROP, self.item.properties)
 
-        links = self.item.get_links(scientific.CITE_AS)
+        links = self.item.get_links(ScientificRelType.CITE_AS)
         doi_urls = [link.get_href() for link in links]
         expected = [PUB1_DOI_URL]
         self.assertCountEqual(expected, doi_urls)
@@ -93,9 +122,9 @@ class ItemScientificExtensionTest(unittest.TestCase):
     def test_publications(self) -> None:
         ScientificExtension.ext(self.item).apply(publications=PUBLICATIONS)
         self.assertEqual(PUBLICATIONS, ScientificExtension.ext(self.item).publications)
-        self.assertIn(scientific.PUBLICATIONS, self.item.properties)
+        self.assertIn(scientific.PUBLICATIONS_PROP, self.item.properties)
 
-        links = self.item.get_links(scientific.CITE_AS)
+        links = self.item.get_links(ScientificRelType.CITE_AS)
         doi_urls = [link.get_href() for link in links]
         expected = [PUB1_DOI_URL, PUB2_DOI_URL]
         self.assertCountEqual(expected, doi_urls)
@@ -106,7 +135,7 @@ class ItemScientificExtensionTest(unittest.TestCase):
         ScientificExtension.ext(self.item).apply(DOI, publications=publications)
         ScientificExtension.ext(self.item).remove_publication(publications[0])
         self.assertFalse(ScientificExtension.ext(self.item).publications)
-        links = self.item.get_links(scientific.CITE_AS)
+        links = self.item.get_links(ScientificRelType.CITE_AS)
         self.assertEqual(1, len(links))
         self.assertEqual(DOI_URL, links[0].target)
         self.item.validate()
@@ -116,7 +145,7 @@ class ItemScientificExtensionTest(unittest.TestCase):
         ScientificExtension.ext(self.item).apply(DOI, publications=publications)
         ScientificExtension.ext(self.item).remove_publication()
         self.assertFalse(ScientificExtension.ext(self.item).publications)
-        links = self.item.get_links(scientific.CITE_AS)
+        links = self.item.get_links(ScientificRelType.CITE_AS)
         self.assertEqual(1, len(links))
         self.assertEqual(DOI_URL, links[0].target)
         self.item.validate()
@@ -128,7 +157,7 @@ class ItemScientificExtensionTest(unittest.TestCase):
         self.assertEqual(
             [PUBLICATIONS[1]], ScientificExtension.ext(self.item).publications
         )
-        links = self.item.get_links(scientific.CITE_AS)
+        links = self.item.get_links(ScientificRelType.CITE_AS)
         self.assertEqual(2, len(links))
         self.assertEqual(DOI_URL, links[0].target)
         self.assertEqual(PUB2_DOI_URL, links[1].target)
@@ -136,7 +165,7 @@ class ItemScientificExtensionTest(unittest.TestCase):
 
         ScientificExtension.ext(self.item).remove_publication(PUBLICATIONS[1])
         self.assertFalse(ScientificExtension.ext(self.item).publications)
-        links = self.item.get_links(scientific.CITE_AS)
+        links = self.item.get_links(ScientificRelType.CITE_AS)
         self.assertEqual(1, len(links))
         self.assertEqual(DOI_URL, links[0].target)
         self.item.validate()
@@ -148,13 +177,13 @@ class ItemScientificExtensionTest(unittest.TestCase):
         self.assertEqual(
             [PUBLICATIONS[0]], ScientificExtension.ext(self.item).publications
         )
-        links = self.item.get_links(scientific.CITE_AS)
+        links = self.item.get_links(ScientificRelType.CITE_AS)
         self.assertEqual(2, len(links))
         self.assertEqual(PUB1_DOI_URL, links[1].target)
 
         self.item.validate()
         ScientificExtension.ext(self.item).remove_publication(PUBLICATIONS[0])
-        links = self.item.get_links(scientific.CITE_AS)
+        links = self.item.get_links(ScientificRelType.CITE_AS)
         self.assertEqual(1, len(links))
         self.assertEqual(DOI_URL, links[0].target)
         self.item.validate()
@@ -163,7 +192,7 @@ class ItemScientificExtensionTest(unittest.TestCase):
         ScientificExtension.ext(self.item).apply(DOI, publications=PUBLICATIONS)
         ScientificExtension.ext(self.item).remove_publication()
         self.assertFalse(ScientificExtension.ext(self.item).publications)
-        links = self.item.get_links(scientific.CITE_AS)
+        links = self.item.get_links(ScientificRelType.CITE_AS)
         self.assertEqual(1, len(links))
         self.assertEqual(DOI_URL, links[0].target)
         self.item.validate()
@@ -172,7 +201,7 @@ class ItemScientificExtensionTest(unittest.TestCase):
         ScientificExtension.ext(self.item).apply(DOI)
         ScientificExtension.ext(self.item).remove_publication()
         self.assertFalse(ScientificExtension.ext(self.item).publications)
-        links = self.item.get_links(scientific.CITE_AS)
+        links = self.item.get_links(ScientificRelType.CITE_AS)
         self.assertEqual(1, len(links))
         self.assertEqual(DOI_URL, links[0].target)
         self.item.validate()
@@ -205,8 +234,8 @@ class CollectionScientificExtensionTest(unittest.TestCase):
     def test_doi(self) -> None:
         ScientificExtension.ext(self.collection).apply(DOI)
         self.assertEqual(DOI, ScientificExtension.ext(self.collection).doi)
-        self.assertIn(scientific.DOI, self.collection.extra_fields)
-        link = self.collection.get_links(scientific.CITE_AS)[0]
+        self.assertIn(scientific.DOI_PROP, self.collection.extra_fields)
+        link = self.collection.get_links(ScientificRelType.CITE_AS)[0]
         self.assertEqual(DOI_URL, link.get_href())
         self.collection.validate()
 
@@ -214,21 +243,21 @@ class CollectionScientificExtensionTest(unittest.TestCase):
 
         # Same doi.
         ScientificExtension.ext(self.collection).doi = DOI
-        self.assertEqual(1, len(self.collection.get_links(scientific.CITE_AS)))
+        self.assertEqual(1, len(self.collection.get_links(ScientificRelType.CITE_AS)))
         self.collection.validate()
 
         # Different doi.
         ScientificExtension.ext(self.collection).doi = PUB1_DOI
-        self.assertEqual(1, len(self.collection.get_links(scientific.CITE_AS)))
-        link = self.collection.get_links(scientific.CITE_AS)[0]
+        self.assertEqual(1, len(self.collection.get_links(ScientificRelType.CITE_AS)))
+        link = self.collection.get_links(ScientificRelType.CITE_AS)[0]
         self.assertEqual(PUB1_DOI_URL, link.get_href())
         self.collection.validate()
 
     def test_citation(self) -> None:
         ScientificExtension.ext(self.collection).apply(citation=CITATION)
         self.assertEqual(CITATION, ScientificExtension.ext(self.collection).citation)
-        self.assertIn(scientific.CITATION, self.collection.extra_fields)
-        self.assertFalse(self.collection.get_links(scientific.CITE_AS))
+        self.assertIn(scientific.CITATION_PROP, self.collection.extra_fields)
+        self.assertFalse(self.collection.get_links(ScientificRelType.CITE_AS))
         self.collection.validate()
 
     def test_publications_one(self) -> None:
@@ -237,9 +266,9 @@ class CollectionScientificExtensionTest(unittest.TestCase):
         self.assertEqual(
             publications, ScientificExtension.ext(self.collection).publications
         )
-        self.assertIn(scientific.PUBLICATIONS, self.collection.extra_fields)
+        self.assertIn(scientific.PUBLICATIONS_PROP, self.collection.extra_fields)
 
-        links = self.collection.get_links(scientific.CITE_AS)
+        links = self.collection.get_links(ScientificRelType.CITE_AS)
         doi_urls = [link.get_href() for link in links]
         expected = [PUB1_DOI_URL]
         self.assertCountEqual(expected, doi_urls)
@@ -251,9 +280,9 @@ class CollectionScientificExtensionTest(unittest.TestCase):
         self.assertEqual(
             PUBLICATIONS, ScientificExtension.ext(self.collection).publications
         )
-        self.assertIn(scientific.PUBLICATIONS, self.collection.extra_fields)
+        self.assertIn(scientific.PUBLICATIONS_PROP, self.collection.extra_fields)
 
-        links = self.collection.get_links(scientific.CITE_AS)
+        links = self.collection.get_links(ScientificRelType.CITE_AS)
         doi_urls = [link.get_href() for link in links]
         expected = [PUB1_DOI_URL, PUB2_DOI_URL]
         self.assertCountEqual(expected, doi_urls)
@@ -265,7 +294,7 @@ class CollectionScientificExtensionTest(unittest.TestCase):
         ScientificExtension.ext(self.collection).apply(DOI, publications=publications)
         ScientificExtension.ext(self.collection).remove_publication(publications[0])
         self.assertFalse(ScientificExtension.ext(self.collection).publications)
-        links = self.collection.get_links(scientific.CITE_AS)
+        links = self.collection.get_links(ScientificRelType.CITE_AS)
         self.assertEqual(1, len(links))
         self.assertEqual(DOI_URL, links[0].target)
         self.collection.validate()
@@ -275,7 +304,7 @@ class CollectionScientificExtensionTest(unittest.TestCase):
         ScientificExtension.ext(self.collection).apply(DOI, publications=publications)
         ScientificExtension.ext(self.collection).remove_publication()
         self.assertFalse(ScientificExtension.ext(self.collection).publications)
-        links = self.collection.get_links(scientific.CITE_AS)
+        links = self.collection.get_links(ScientificRelType.CITE_AS)
         self.assertEqual(1, len(links))
         self.assertEqual(DOI_URL, links[0].target)
         self.collection.validate()
@@ -287,7 +316,7 @@ class CollectionScientificExtensionTest(unittest.TestCase):
         self.assertEqual(
             [PUBLICATIONS[1]], ScientificExtension.ext(self.collection).publications
         )
-        links = self.collection.get_links(scientific.CITE_AS)
+        links = self.collection.get_links(ScientificRelType.CITE_AS)
         self.assertEqual(2, len(links))
         self.assertEqual(DOI_URL, links[0].target)
         self.assertEqual(PUB2_DOI_URL, links[1].target)
@@ -295,7 +324,7 @@ class CollectionScientificExtensionTest(unittest.TestCase):
 
         ScientificExtension.ext(self.collection).remove_publication(PUBLICATIONS[1])
         self.assertFalse(ScientificExtension.ext(self.collection).publications)
-        links = self.collection.get_links(scientific.CITE_AS)
+        links = self.collection.get_links(ScientificRelType.CITE_AS)
         self.assertEqual(1, len(links))
         self.assertEqual(DOI_URL, links[0].target)
         self.collection.validate()
@@ -307,13 +336,13 @@ class CollectionScientificExtensionTest(unittest.TestCase):
         self.assertEqual(
             [PUBLICATIONS[0]], ScientificExtension.ext(self.collection).publications
         )
-        links = self.collection.get_links(scientific.CITE_AS)
+        links = self.collection.get_links(ScientificRelType.CITE_AS)
         self.assertEqual(2, len(links))
         self.assertEqual(PUB1_DOI_URL, links[1].target)
 
         self.collection.validate()
         ScientificExtension.ext(self.collection).remove_publication(PUBLICATIONS[0])
-        links = self.collection.get_links(scientific.CITE_AS)
+        links = self.collection.get_links(ScientificRelType.CITE_AS)
         self.assertEqual(1, len(links))
         self.assertEqual(DOI_URL, links[0].target)
         self.collection.validate()
@@ -322,7 +351,7 @@ class CollectionScientificExtensionTest(unittest.TestCase):
         ScientificExtension.ext(self.collection).apply(DOI, publications=PUBLICATIONS)
         ScientificExtension.ext(self.collection).remove_publication()
         self.assertFalse(ScientificExtension.ext(self.collection).publications)
-        links = self.collection.get_links(scientific.CITE_AS)
+        links = self.collection.get_links(ScientificRelType.CITE_AS)
         self.assertEqual(1, len(links))
         self.assertEqual(DOI_URL, links[0].target)
         self.collection.validate()
@@ -331,10 +360,49 @@ class CollectionScientificExtensionTest(unittest.TestCase):
         ScientificExtension.ext(self.collection).apply(DOI)
         ScientificExtension.ext(self.collection).remove_publication()
         self.assertFalse(ScientificExtension.ext(self.collection).publications)
-        links = self.collection.get_links(scientific.CITE_AS)
+        links = self.collection.get_links(ScientificRelType.CITE_AS)
         self.assertEqual(1, len(links))
         self.assertEqual(DOI_URL, links[0].target)
         self.collection.validate()
+
+
+class SummariesScientificTest(unittest.TestCase):
+    def setUp(self) -> None:
+        super().setUp()
+        summaries = Summaries(
+            summaries={"sci:citation": [CITATION], "sci:doi": [PUB1_DOI, PUB2_DOI]}
+        )
+        self.collection = make_collection()
+        self.collection.summaries = summaries
+
+    def test_get_citation_summaries(self) -> None:
+        citations = ScientificExtension.summaries(self.collection).citation
+
+        assert citations is not None
+        self.assertListEqual([CITATION], citations)
+
+    def test_set_citation_summaries(self) -> None:
+        collection = self.collection.clone()
+        sci_summaries = ScientificExtension.summaries(collection)
+
+        sci_summaries.citation = None
+        self.assertIsNone(sci_summaries.citation)
+
+    def test_get_doi_summaries(self) -> None:
+        dois = ScientificExtension.summaries(self.collection).doi
+
+        assert dois is not None
+        self.assertListEqual([PUB1_DOI, PUB2_DOI], dois)
+
+    def test_set_doi_summaries(self) -> None:
+        collection = self.collection.clone()
+        sci_summaries = ScientificExtension.summaries(collection)
+
+        sci_summaries.doi = [PUB2_DOI]
+        new_dois = ScientificExtension.summaries(self.collection).doi
+
+        assert new_dois is not None
+        self.assertListEqual([PUB2_DOI], new_dois)
 
 
 if __name__ == "__main__":